+++ src/wav.cpp
@@ -68,7 +68,7 @@ rsrc * wav_func64__sub0(NC_STACK_wav *obj, stack_vals *stak, const char *filname
                 {
                     stack_vals stk[3];
 
-                    stk[0].set(NC_STACK_sample::SMPL_ATT_LEN, sbchunk.SubchunkSize);
+                    stk[0].set(NC_STACK_sample::SMPL_ATT_LEN, (int)sbchunk.SubchunkSize);
                     stk[1].set(NC_STACK_sample::SMPL_ATT_TYPE, 1);
                     stk[2].nextStack(stak);
